src/runtime/requirements.txt
aws-lambda-powertools~=2.25.1
aws-lambda-powertools[tracer]~=2.25.1
psycopg2-binary~=2.9.5
